fix: Fix deadlock when child gets kill -9 sig

+++ fastembed/parallel_processor.py
@@ -83,9 +83,7 @@ def _worker(
 
 
 class ParallelWorkerPool:
-    def __init__(
-        self, num_workers: int, worker: Type[Worker], start_method: Optional[str] = None
-    ):
+    def __init__(self, num_workers: int, worker: Type[Worker], start_method: Optional[str] = None):
         self.worker_class = worker
         self.num_workers = num_workers
         self.input_queue: Optional[Queue] = None
@@ -120,9 +118,7 @@ class ParallelWorkerPool:
             process.start()
             self.processes.append(process)
 
-    def ordered_map(
-        self, stream: Iterable[Any], *args: Any, **kwargs: Any
-    ) -> Iterable[Any]:
+    def ordered_map(self, stream: Iterable[Any], *args: Any, **kwargs: Any) -> Iterable[Any]:
         buffer = defaultdict(Any)
         next_expected = 0
 
@@ -144,6 +140,7 @@ class ParallelWorkerPool:
             pushed = 0
             read = 0
             for idx, item in enumerate(stream):
+                self.check_worker_health()
                 if pushed - read < self.queue_size:
                     try:
                         out_item = self.output_queue.get_nowait()
@@ -170,6 +167,7 @@ class ParallelWorkerPool:
                 self.input_queue.put(QueueSignals.stop)
 
             while read < pushed:
+                self.check_worker_health()
                 out_item = self.output_queue.get(timeout=processing_timeout)
                 if out_item == QueueSignals.error:
                     self.join_or_terminate()
@@ -181,6 +179,17 @@ class ParallelWorkerPool:
             assert self.output_queue is not None, "Output queue is None"
             self.input_queue.close()
             self.output_queue.close()
+
+    def check_worker_health(self) -> None:
+        """
+        Checks if any worker process has terminated unexpectedly
+        """
+        for process in self.processes:
+            if not process.is_alive() and process.exitcode != 0:
+                self.join_or_terminate()
+                raise RuntimeError(
+                    f"Worker (PID: {process.pid} terminated unexpectedly with code {process.exitcode}"
+                )
 
     def join_or_terminate(self, timeout: Optional[int] = 1) -> None:
         """
@@ -210,4 +219,6 @@ class ParallelWorkerPool:
         https://eli.thegreenplace.net/2009/06/12/safely-using-destructors-in-python/.
         """
         for process in self.processes:
-            process.terminate()
+            if process.is_alive():
+                process.terminate()
+        self.processes.clear()
